Bench: 7231124

+++ Halogen/src/Position.h
@@ -47,7 +47,7 @@ public:
 
 	Network net;
 
-	float GetEvaluation();
+	int16_t GetEvaluation();
 
 private:
 	uint64_t key;
@@ -56,7 +56,7 @@ private:
 	uint64_t GenerateZobristKey() const;
 	uint64_t IncrementZobristKey(Move move);	
 
-	std::vector<float> GetInputLayer() const;
+	std::vector<int16_t> GetInputLayer() const;
 	std::vector<deltaPoint>& CalculateMoveDelta(Move move);				//A vector which calculates the CHANGE in each input paramiter
 
 	static size_t modifier(size_t index);								//no inputs for pawns on front or back rank for neural net: we need to modify zobrist-like indexes
